requirements_model.txt
pandas>=1.3
numpy>=1.21
pyyaml>=6.0
scikit-learn>=1.0
xgboost>=1.7
lightgbm>=3.3
yfinance>=0.2
akshare>=1.13
